requirements.txt
fastapi
uvicorn[standard]
hydra-core
pydantic
python-multipart
python-dotenv
supabase
email-validator
SQLAlchemy